+++ src/gcovr/formats/gcov/parser/text.py
@@ -263,6 +263,7 @@ def parse_coverage(
     ignore_parse_errors: Optional[set[str]],
     suspicious_hits_threshold: int = SUSPICIOUS_COUNTER,
     activate_trace_logging: bool = False,
+    use_existing_files: bool = False,
 ) -> tuple[FileCoverage, list[str]]:
     """
     Extract coverage data from a gcov report.
@@ -305,6 +306,30 @@ def parse_coverage(
         except Exception as ex:  # pylint: disable=broad-except
             lines_with_errors.append((raw_line, ex))
 
+    missing_function_lines = not any(
+        isinstance(line, _FunctionLine) for line, _ in tokenized_lines
+    )
+
+    if (
+        use_existing_files
+        # Only check for missing function lines if we have any non-metadata lines.
+        and any(not isinstance(line, _MetadataLine) for line, _ in tokenized_lines)
+        and missing_function_lines
+    ):
+        files_for_message = (
+            "\n   ".join(
+                " -> ".join(file_tuple) for file_tuple in sorted(data_filename)
+            )
+            if isinstance(data_filename, set)
+            else data_filename
+        )
+        LOGGER.warning(
+            f"No function line found in gcov file:\n   {files_for_message}\n"
+            "This may indicate that the file was generated without "
+            "the proper gcov options (especially --branch-probabilities)?\n"
+            "See <https://gcovr.com/en/stable/faq.html#which-options-are-used-for-calling-gcov>.",
+        )
+
     if (
         "negative_hits.warn_once_per_file" in persistent_states
         and persistent_states["negative_hits.warn_once_per_file"] > 1
@@ -325,6 +350,8 @@ def parse_coverage(
 
     filecov = FileCoverage(data_filename, filename=filename)
     state = _ParserState()
+    if use_existing_files and missing_function_lines:
+        state = state._replace(function_name="<unknown>")
     for line, raw_line in tokenized_lines:
         try:
             state = _gather_coverage_from_line(
@@ -611,7 +638,10 @@ def _report_lines_with_errors(
 
     lines_output = "\n\t  ".join(lines)
     LOGGER.warning(
-        "Unrecognized GCOV output for %s\n\t  %s\n\tThis is indicative of a gcov output parse error.\n\tPlease report this to the gcovr developers\n\tat <https://github.com/gcovr/gcovr/issues>.",
+        "Unrecognized GCOV output for %s\n\t  %s\n"
+        "\tThis is indicative of a gcov output parse error.\n"
+        "\tPlease report this to the gcovr developers\n"
+        "\tat <https://github.com/gcovr/gcovr/issues>.",
         filename,
         lines_output,
     )