+++ src/gcovr/formats/gcov/workers.py
@@ -128,6 +128,9 @@ class Workers:
     1
     """
 
+    class WorkerThreadException(RuntimeError):
+        """Exception raised when a worker thread fails."""
+
     def __init__(self, number: int, context: Callable[[], dict[str, Any]]) -> None:
         if number <= 0:
             number = max(1, cpu_count() + number)
@@ -200,11 +203,10 @@ class Workers:
                 w.join(timeout=1)
         self.workers = []
 
-        for traceback in self.exceptions:
-            LOGGER.error(traceback)
-
         if self.exceptions:
-            raise RuntimeError(
+            for traceback in self.exceptions:
+                LOGGER.error(traceback)
+            raise self.WorkerThreadException(
                 "Worker thread raised exception, workers canceled."
             ) from None
         return self.contexts